\name{grp}
 \docType{data}
 \alias{grp}
 \title{Group of the individuals.}
 \description{
 This is a vector of the group of the simulated profiles contained in the matrix "dat" in
 the k-TSP package. The groups are "healthy" and "diseased".
 }
 \usage{grp}
 \format{vector}
 \keyword{datasets}