 \name{ktspdata} \docType{data} \alias{ktspdata} \title{Simulated dataset of gene expressions in a matrix form.} \description{ This dataset can be used to illustrate the functions contained in the package "ktspair". The gene expressions are simulated from poisson random variables, where the means come from a normal distribution with different parameters for the group of healthy and diseased people. The first 25 measures are healthy  people (represented by 0 in the group vector) and the 25 last measures are diseased people (represented by 1 in the group vector). } \usage{dat} \keyword{datasets}